backend/requirements.txt
autopep8==2.0.4
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
google-api-core==2.12.0
google-auth==2.23.3
google-cloud-vision==3.4.5
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
idna==3.4
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
requests==2.31.0
rsa==4.9
urllib3==2.0.7
